Changed queryset to only filter on 'list' action.
This is for more predictable http response codes. Filtering the
queryset for all actions led to 403s and 404s, depending on if the user
could see the resource at all. If a user could see a resource but not
edit it, it would return 403; otherwise, if the user couldn't see the
video and also couldn't edit it, it would return 404. Now we just use
Video.objects.all() which ensures a 404 if the resource doesn't exist, or
403 if it does but it can't be accessed.